first login show tutorial ok
