+++ MapConvertQuestion/AppDelegate.swift
@@ -16,21 +16,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         
-        let defaults = UserDefaults.standard
-        defaults.register(defaults: ["FirstLaunch" : true])
-        
-        if defaults.bool(forKey: "FirstLaunch") == true {
-            UserDefaults.standard.set(false, forKey: "FirstLaunch")
-            let new_uuid = NSUUID().uuidString
-            UserDefaults.standard.set(new_uuid, forKey: "uuid")
-             print("new!! \n \(new_uuid) \n\n")
-            print("初回のログインです")
-            self.createUserData(uuid:new_uuid)
-        }else{
-            UserDefaults.standard.set(false, forKey: "FirstLaunch")
-            print("２回目以降のログインです")
-        }
-        
         let migSchemaVersion: UInt64 = 16
         let config = Realm.Configuration(
             schemaVersion: migSchemaVersion,
@@ -42,22 +27,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         return true
     }
     
-    private func createUserData(uuid:String){
-        do {
-            let realm = try Realm()
-            let user = User(value: [
-                "uuid": uuid
-            ])
-            try! realm.write {
-                realm.add(user)
-                print("成功", user)
-            }
-        } catch {
-            print("\(error)")
-            print("エラーだよ")
-        }
-    }
-
     // MARK: UISceneSession Lifecycle
 
     @available(iOS 13.0, *)
